Safari-609.1.11.1/LayoutTests/fast/text/variations/advances-expected.txt

PASS specifiedregular.offsetWidth is not wide.offsetWidth
PASS specifiedregular.offsetWidth is not narrow.offsetWidth
PASS narrow.offsetWidth is not wide.offsetWidth
PASS successfullyParsed is true

TEST COMPLETE
ham
ham
ham
ham
